% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDtoR.R
\name{MDtoR}
\alias{MDtoR}
\title{Safely read file contents}
\usage{
MDtoR(file, export = FALSE)
}
\arguments{
\item{file}{Path to file}

\item{export}{Export extracted R code to a file (default: FALSE).}
}
\value{
Character vector of file content
}
\description{
Safely read file contents
}
