% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vault.R
\name{read_vault}
\alias{read_vault}
\title{Reads all notes from Obsidian vault}
\usage{
read_vault(path)
}
\arguments{
\item{path}{Path to the Obsidian vault

Reads only .md and .canvas files}
}
\description{
Reads all notes from Obsidian vault
}
