% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{safe_read_file}
\alias{safe_read_file}
\title{Safely read file contents}
\usage{
safe_read_file(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Character vector of file content
}
\description{
Safely read file contents
}
