% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{utils_extract_yaml}
\alias{utils_extract_yaml}
\title{Extract YAML header from content}
\usage{
utils_extract_yaml(content)
}
\arguments{
\item{content}{Character vector of file content}
}
\value{
List with yaml and content components
}
\description{
Extract YAML header from content
}
