% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{utils_process_rmd_file}
\alias{utils_process_rmd_file}
\title{Process Rmd file}
\usage{
utils_process_rmd_file(file_path, preserve_yaml)
}
\arguments{
\item{file_path}{Path to Rmd file}

\item{preserve_yaml}{Boolean to preserve YAML}
}
\value{
List with content and yaml components
}
\description{
Process Rmd file
}
